pyproject.toml
[project]
name = "ai_agents_01_call_llm"
version = "0.1.0"
description = "LLM API + math tool roundtrip"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.40.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
llm-tool = "ai_agents_01_call_llm.main:main"

# --- packaging for src/ layout ---
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
